requirement.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic
pypdf
pandas
httpx
